RAG_Based_ChatBot/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-dotenv
pydantic
pypdf
numpy
tiktoken